[refactoring] logic을 담고있는 부분 전부 domain 패키지 이동
