+++ src/main/java/racing/domain/Cars.java
@@ -1,10 +1,12 @@
-package racing.model;
+package racing.domain;
 
 import racing.strategy.MoveStrategy;
 
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Cars {
 
@@ -29,5 +31,13 @@ public class Cars {
         for (Car car : carList) {
             car.movable(moveStrategy);
         }
+    }
+
+    public Integer findMaxPosition(Cars cars) {
+        return cars.getCarList().stream().map(i -> i.getPosition().getPosition()).max(Comparator.comparing(x -> x)).get();
+    }
+
+    public List<Car> findWinnerList(Cars cars) {
+        return cars.getCarList().stream().filter(i -> i.getPosition().getPosition() >= findMaxPosition(cars)).collect(Collectors.toList());
     }
 }
