+++ src/main/java/racing/domain/InputBuilder.java
@@ -1,4 +1,4 @@
-package racing.service;
+package racing.domain;
 
 import java.util.Arrays;
 import java.util.List;
